Extracts LocustOIDCClient class
