+++ matrix_locust/nio/locust_client.py
@@ -208,36 +208,6 @@ class LocustClient(Client):
         # Hacky way to allow unpacking a tuple return value
         return (*return_items,)
 
-    def _clean_response_dict(self, data):
-        """Clean response dictionary by replacing None values with empty objects/lists where expected."""
-        if not isinstance(data, dict):
-            return data
-
-        cleaned = {}
-        for key, value in data.items():
-            if value is None:
-                if key in ['rooms', 'account_data', 'presence', 'device_lists']:
-                    cleaned[key] = {}
-                elif key in ['events', 'chunk', 'prev_content', 'state_events']:
-                    cleaned[key] = []
-                else:
-                    cleaned[key] = value
-            elif isinstance(value, dict):
-                cleaned[key] = self._clean_response_dict(value)
-            elif isinstance(value, list):
-                cleaned[key] = [self._clean_response_dict(item) if isinstance(item, dict) else item for item in value]
-            else:
-                cleaned[key] = value
-
-        if 'start' in data and 'chunk' not in cleaned:
-            cleaned['chunk'] = []
-        if 'start' in data and 'end' not in cleaned:
-            cleaned['end'] = cleaned.get('start', '')
-
-        if 'next_batch' in data and 'rooms' not in cleaned:
-            cleaned['rooms'] = {'join': {}, 'invite': {}, 'leave': {}}
-
-        return cleaned
 
     def _send(
         self,
@@ -258,28 +228,11 @@ class LocustClient(Client):
 
         # Send request and update internal state of the object with the response
         # logging.info("[%s] Making API call to %s" % (self.user, url))
-        with self.locust_user.client.request(
-            method, url, headers=headers, json=body, name=name, catch_response=True
-        ) as resp:
-            try:
-                response_json = resp.json() if resp.content else {}
-                if response_json is None:
-                    response_json = {}
-                elif isinstance(response_json, dict):
-                    response_json = self._clean_response_dict(response_json)
-
-                matrix_response = response.from_dict(response_json, *response_data)
-                self.receive_response(matrix_response)
-                self.run_response_callbacks([matrix_response])
-                resp.success()
-                return matrix_response
-            except Exception as e:
-                logging.error(f"Error processing Matrix response for {url}: {e}")
-                logging.error(f"Raw response: {resp.json() if resp.content else 'No content'}")
-                from nio.responses import ErrorResponse
-                error_response = ErrorResponse("Response parsing failed", "M_UNKNOWN", resp.status_code)
-                resp.failure(f"Response parsing failed: {e}")
-                return error_response
+        with self.locust_user.rest(method, url, headers=headers, json=body, name=name) as resp:
+            matrix_response = response.from_dict(resp.js, *response_data)
+            self.receive_response(matrix_response)
+            self.run_response_callbacks([matrix_response])
+            return matrix_response
 
     def add_response_callback(
         self,
@@ -365,364 +318,6 @@ class LocustClient(Client):
 
         return response
 
-    def login_oidc(
-        self,
-        oidc_issuer: str,
-        client_id: str = "matrix-locust",
-        device_name: Optional[str] = "",
-        redirect_uri: str = "http://localhost:8080/callback",
-        username: str = None,
-        password: str = None,
-    ) -> Union[LoginResponse, LoginError]:
-        """Login to the homeserver using real OIDC authentication.
-
-        This method implements the actual Matrix SSO/OIDC flow including:
-        1. Initiating SSO redirect with the Matrix homeserver
-        2. Handling NitroID authentication form submission
-        3. Following the callback to get login token
-        4. Using the login token to authenticate with Matrix
-
-        Args:
-            oidc_issuer (str): The OIDC issuer URL (e.g., https://id.powerhrg.com).
-            client_id (str): The OIDC client ID.
-            device_name (str): A display name for the device.
-            redirect_uri (str): The redirect URI for OIDC callback.
-            username (str): Username for NitroID login.
-            password (str): Password for NitroID login.
-
-        Returns either a `LoginResponse` if the request was successful or
-        a `LoginError` if there was an error with the request.
-        """
-        try:
-            # Get credentials from client attributes if not provided
-            if username is None:
-                username = getattr(self, "oidc_username", None)
-            if password is None:
-                password = getattr(self, "oidc_password", None)
-
-            if not username or not password:
-                return LoginError(
-                    "OIDC username/password not provided",
-                    status_code="M_OIDC_CREDENTIALS_MISSING",
-                )
-
-            # Step 1: Get the Matrix SSO login URL
-            login_token = self._perform_real_oidc_flow(username, password, redirect_uri)
-
-            if not login_token:
-                return LoginError(
-                    "Failed to obtain login token from OIDC flow",
-                    status_code="M_OIDC_TOKEN_MISSING",
-                )
-
-            # Step 2: Use the login token to authenticate with Matrix
-            method, path, data = self._build_request(
-                Api.login(
-                    self.user,
-                    password=None,
-                    device_name=device_name,
-                    device_id=self.device_id,
-                    token=login_token,
-                )
-            )
-
-            response = self._send(LoginResponse, method, path, data)
-
-            if isinstance(response, LoginResponse):
-                self.matrix_domain = self.user_id.split(":")[-1]
-
-            return response
-
-        except Exception as e:
-            return LoginError(
-                f"OIDC authentication failed: {str(e)}", status_code="M_OIDC_ERROR"
-            )
-
-    def _perform_real_oidc_flow(
-        self, username: str, password: str, redirect_uri: str
-    ) -> Optional[str]:
-        """Perform the actual OIDC authentication flow with NitroID.
-
-        This method implements the real Matrix SSO + OIDC flow:
-        1. Start SSO redirect with Matrix homeserver
-        2. Follow redirects to NitroID
-        3. Submit login form to NitroID
-        4. Follow callback redirects to get login token
-
-        Args:
-            username: NitroID username
-            password: NitroID password
-            redirect_uri: Callback URI for OIDC flow
-
-        Returns:
-            Login token from Matrix callback, or None if authentication failed
-        """
-        max_retries = 3
-
-        for attempt in range(max_retries):
-            session = requests.Session()
-            session.timeout = 30
-
-            # Configure session with proper headers to avoid bot detection
-            session.headers.update(
-                {
-                    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36",
-                    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
-                    "Accept-Language": "en-US,en;q=0.5",
-                    "Accept-Encoding": "gzip, deflate",
-                    "Connection": "keep-alive",
-                }
-            )
-
-            try:
-                print(f"OIDC login attempt {attempt + 1}/{max_retries}")
-
-                # Step 1: Get the Matrix SSO login URL
-                # Use the Matrix SSO redirect endpoint with NitroID IDP hint
-                matrix_base_url = f"{self.locust_user.host}"
-                sso_redirect_url = f"{matrix_base_url}/_matrix/client/v3/login/sso/redirect/oidc-nitroid"
-
-                print(f"Starting SSO flow to {sso_redirect_url}")
-                print(f"Using callback URL: {redirect_uri}")
-                sso_params = {"redirectUrl": redirect_uri}
-                response = session.get(
-                    sso_redirect_url, params=sso_params, allow_redirects=True
-                )
-                response.raise_for_status()
-
-                print(
-                    f"Final page reached: (status: {response.status_code})"
-                )
-
-                # Step 3: We should now be at the NitroID login page
-                # Parse the login form and submit credentials
-                login_token = self._handle_nitroid_login(
-                    session, response, username, password, redirect_uri
-                )
-
-                if login_token:
-                    return login_token
-                else:
-                    print(f"Attempt {attempt + 1} failed to get login token")
-
-            except Exception as e:
-                print(f"Attempt {attempt + 1} failed with exception: {str(e)}")
-                if attempt == max_retries - 1:
-                    print(f"All {max_retries} attempts failed")
-                    return None
-                else:
-                    print(f"Retrying in 2 seconds...")
-                    time.sleep(2)
-            finally:
-                session.close()
-
-        return None
-
-    def _handle_nitroid_login(
-        self,
-        session: requests.Session,
-        response: requests.Response,
-        username: str,
-        password: str,
-        redirect_uri: str,
-    ) -> Optional[str]:
-        """Handle the NitroID login form submission.
-
-        Args:
-            session: requests session maintaining cookies
-            response: response from NitroID login page
-            username: NitroID username
-            password: NitroID password
-            redirect_uri: Callback URI
-
-        Returns:
-            Login token from callback, or None if failed
-        """
-        try:
-
-            # Parse the login form from the NitroID page
-            soup = BeautifulSoup(response.text, "html.parser")
-
-            # Find the login form - try multiple strategies
-            login_form = None
-
-            # Strategy 1: Look for forms with login-related attributes
-            for form in soup.find_all("form"):
-                form_id = form.get("id", "").lower()
-                form_class = " ".join(form.get("class", [])).lower()
-                form_action = form.get("action", "").lower()
-
-                if any(
-                    keyword in form_id + form_class + form_action
-                    for keyword in ["login", "signin", "auth", "credential"]
-                ):
-                    login_form = form
-                    break
-
-            # Strategy 2: Look for forms with password fields
-            if not login_form:
-                for form in soup.find_all("form"):
-                    if form.find("input", {"type": "password"}):
-                        login_form = form
-                        break
-
-            # Strategy 3: Use the first form as fallback
-            if not login_form:
-                login_form = soup.find("form")
-
-            if not login_form:
-                print("Could not find any login form on NitroID page")
-                print(f"Page title: {soup.title.string if soup.title else 'No title'}")
-                print(f"Available forms: {len(soup.find_all('form'))}")
-                print(f"Page preview: {response.text[:500]}...")
-                return None
-
-            form_action = login_form.get("action")
-            if not form_action.startswith("http"):
-                # Relative URL, make it absolute
-                base_url = f"{response.url.split('?')[0].rsplit('/', 1)[0]}"
-                form_action = f"{base_url}/{form_action.lstrip('/')}"
-
-            # Extract any hidden form fields (CSRF tokens, etc.)
-            form_data = {}
-            for input_field in login_form.find_all("input", {"type": "hidden"}):
-                name = input_field.get("name")
-                value = input_field.get("value", "")
-                if name:
-                    form_data[name] = value
-
-            # Add username and password fields
-            # Common field names for username/email
-            username_fields = ["email", "username", "login", "user"]
-            password_fields = ["password", "passwd", "pwd"]
-
-            # Find the actual field names from the form
-            username_field_found = False
-            password_field_found = False
-
-            for input_field in login_form.find_all("input"):
-                field_type = input_field.get("type", "").lower()
-                field_name = input_field.get("name", "").lower()
-                field_id = input_field.get("id", "").lower()
-                field_placeholder = input_field.get("placeholder", "").lower()
-
-                # Enhanced username field detection
-                if (
-                    field_type in ["email", "text"]
-                    or any(uf in field_name for uf in username_fields)
-                    or any(uf in field_id for uf in username_fields)
-                    or any(
-                        uf in field_placeholder for uf in ["email", "username", "user"]
-                    )
-                ):
-                    form_data[input_field.get("name")] = username
-                    username_field_found = True
-
-                # Enhanced password field detection
-                elif (
-                    field_type == "password"
-                    or any(pf in field_name for pf in password_fields)
-                    or any(pf in field_id for pf in password_fields)
-                    or "password" in field_placeholder
-                ):
-                    form_data[input_field.get("name")] = password
-                    password_field_found = True
-
-            if not username_field_found:
-                print("WARNING: Could not identify username field, trying fallback...")
-                # Fallback: look for any text input that's not hidden
-                for input_field in login_form.find_all("input"):
-                    if (
-                        input_field.get("type", "").lower() in ["text", "email", ""]
-                        and input_field.get("type", "").lower() != "hidden"
-                    ):
-                        form_data[input_field.get("name")] = username
-                        break
-
-            if not password_field_found:
-                print("WARNING: Could not identify password field")
-                # This is more critical - we should see a password field
-
-
-            # Submit the login form
-            login_response = session.post(
-                form_action, data=form_data, allow_redirects=True
-            )
-            login_response.raise_for_status()
-
-            # Step 4: Follow redirects to get back to Matrix with login token
-            # The callback should contain a login token in the URL
-            final_url = login_response.url
-
-            # Extract login token from the callback URL and response
-            parsed_url = urllib.parse.urlparse(final_url)
-            query_params = urllib.parse.parse_qs(parsed_url.query)
-
-            login_token = None
-
-            # Strategy 1: Check URL parameters
-            if "loginToken" in query_params:
-                login_token = query_params["loginToken"][0]
-
-            # Strategy 2: Check for common token parameter variations
-            elif "token" in query_params:
-                login_token = query_params["token"][0]
-            elif "access_token" in query_params:
-                login_token = query_params["access_token"][0]
-
-            # Strategy 3: Check response headers
-            if not login_token:
-                auth_header = login_response.headers.get("Authorization", "")
-                if auth_header.startswith("Bearer "):
-                    login_token = auth_header[7:]
-                elif "X-Login-Token" in login_response.headers:
-                    login_token = login_response.headers["X-Login-Token"]
-
-            # Strategy 4: Check response body with multiple patterns
-            if not login_token:
-                patterns = [
-                    r'loginToken["\']?\s*[:=]\s*["\']?([^"\'&\s<>]+)',
-                    r'token["\']?\s*[:=]\s*["\']?([^"\'&\s<>]+)',
-                    r'access_token["\']?\s*[:=]\s*["\']?([^"\'&\s<>]+)',
-                    r'["\']loginToken["\']\s*:\s*["\']([^"\']+)',
-                    r'window\.location\.href\s*=\s*["\'][^"\']*[?&]loginToken=([^"\'&]+)',
-                ]
-
-                for pattern in patterns:
-                    token_match = re.search(pattern, login_response.text, re.IGNORECASE)
-                    if token_match:
-                        login_token = token_match.group(1)
-                        break
-
-            # Strategy 5: Look for JavaScript redirects or meta refresh
-            if not login_token:
-                # Check for meta refresh with token
-                meta_match = re.search(
-                    r'<meta[^>]+refresh[^>]+url=([^"\'>\s]+)',
-                    login_response.text,
-                    re.IGNORECASE,
-                )
-                if meta_match:
-                    redirect_url = meta_match.group(1)
-                    redirect_params = urllib.parse.parse_qs(
-                        urllib.parse.urlparse(redirect_url).query
-                    )
-                    if "loginToken" in redirect_params:
-                        login_token = redirect_params["loginToken"][0]
-
-            if login_token:
-                return login_token
-            else:
-                print("Could not extract login token from callback")
-                print(f"Final URL: {final_url}")
-                print(f"URL query params: {query_params}")
-                print(f"Response headers: {dict(login_response.headers)}")
-                print(f"Response text preview: {login_response.text[:1000]}...")
-                return None
-
-        except Exception as e:
-            print(f"Error handling NitroID login: {str(e)}")
-            return None
 
     @logged_in
     def logout(self, all_devices: bool = False) -> Union[LogoutResponse, LogoutError]:
@@ -778,17 +373,17 @@ class LocustClient(Client):
 
         data = json.loads(data)
 
-        with self.locust_user.client.request(method, path, json=data, catch_response=True) as response1:
+        with self.locust_user.rest(method, path, json=data) as response1:
             if response1.status_code == HTTPStatus.OK:  # 200
                 logging.info("User [%s] Success!  Didn't even need UIAA!", self.user)
-                self.user_id = response1.json().get("user_id", None)
-                self.access_token = response1.json().get("access_token", None)
+                self.user_id = response1.js.get("user_id", None)
+                self.access_token = response1.js.get("access_token", None)
                 self.matrix_domain = self.user_id.split(":")[-1]
                 if self.user_id is None or self.access_token is None:
                     logging.error(
                         "User [%s] Failed to parse /register response!\nResponse: %s",
                         self.user,
-                        response1.json(),
+                        response1.js,
                     )
                     return
                 self.locust_user.update_tokens()
@@ -796,17 +391,17 @@ class LocustClient(Client):
                 # Not an error, unauthorized requests are apart of the registration-flow
                 response1.success()
 
-                flows = response1.json().get("flows", None)
+                flows = response1.js.get("flows", None)
                 if flows is None:
                     logging.error(
                         "User [%s] No UIAA flows for /register\nResponse: %s",
                         self.user,
-                        response1.json(),
+                        response1.js,
                     )
                     self.locust_user.environment.runner.quit()
                     return
 
-                session_id = response1.json().get("session", None)
+                session_id = response1.js.get("session", None)
                 if session_id is None:
                     logging.info(
                         "User [%s] No session ID provided by server for /register",
@@ -833,17 +428,15 @@ class LocustClient(Client):
                             elif stage == "m.login.registration_token":
                                 data["auth"]["token"] = token
 
-                            with self.locust_user.client.request(
-                                "POST", path, json=data, catch_response=True
-                            ) as response2:
-                                print(response2.json())
+                            with self.locust_user.rest("POST", path, json=data) as response2:
+                                print(response2.js)
                                 if (
                                     response2.status_code == HTTPStatus.OK
                                     or response2.status_code == HTTPStatus.CREATED
                                 ):  # 200 or 201
                                     logging.info("User [%s] Success!", self.user)
-                                    self.user_id = response2.json().get("user_id", None)
-                                    self.access_token = response2.json().get(
+                                    self.user_id = response2.js.get("user_id", None)
+                                    self.access_token = response2.js.get(
                                         "access_token", None
                                     )
                                     self.matrix_domain = self.user_id.split(":")[-1]
@@ -854,7 +447,7 @@ class LocustClient(Client):
                                         logging.error(
                                             "User [%s] Failed to parse /register response!\nResponse: %s",
                                             self.user,
-                                            response2.json(),
+                                            response2.js,
                                         )
                                         return
                                     self.locust_user.update_tokens()
@@ -898,18 +491,16 @@ class LocustClient(Client):
         blind = client.generate_blind()
 
         # Request 1: Empty #####################################################
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json={}, catch_response=True
-        ) as r1:
-            initial_json = r1.json()
-            session_id = r1.json().get("session", None)
+        with self.locust_user.client.request("POST", url, headers=headers, json={}, catch_response=True) as r1:
+            initial_json = r1.js
+            session_id = r1.js.get("session", None)
 
             # print("Got response: ", json.dumps(r1.js, indent=4))
             if r1.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r1.success()
             else:
-                error = r1.json().get("error", "???")
-                errcode = r1.json().get("errcode", "???")
+                error = r1.js.get("error", "???")
+                errcode = r1.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
 
@@ -918,17 +509,15 @@ class LocustClient(Client):
             "username": self.user,
             "auth": {"type": "m.login.terms", "session": session_id},
         }
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json=body, catch_response=True
-        ) as r3:
-            completed = r3.json().get("completed", [])
+        with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r3:
+            completed = r3.js.get("completed", [])
 
             # print("Got response: ", json.dumps(r3.js, indent=4))
             if r3.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r3.success()
             else:
-                error = r3.json().get("error", "???")
-                errcode = r3.json().get("errcode", "???")
+                error = r3.js.get("error", "???")
+                errcode = r3.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
 
@@ -940,17 +529,15 @@ class LocustClient(Client):
                 "username": self.user,
             }
         }
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json=body, catch_response=True
-        ) as r4:
-            completed = r4.json().get("completed", [])
+        with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r4:
+            completed = r4.js.get("completed", [])
 
             # print("Got response: ", json.dumps(r4.js, indent=4))
             if r4.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r4.success()
             else:
-                error = r4.json().get("error", "???")
-                errcode = r4.json().get("errcode", "???")
+                error = r4.js.get("error", "???")
+                errcode = r4.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
 
@@ -968,18 +555,16 @@ class LocustClient(Client):
             }
         }
         bs_speke_params = None
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json=body, catch_response=True
-        ) as r6:
-            bs_speke_params = r6.json()
-            completed = r6.json().get("completed", [])
-            r6_params = r6.json().get("params", {})
+        with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r6:
+            bs_speke_params = r6.js
+            completed = r6.js.get("completed", [])
+            r6_params = r6.js.get("params", {})
 
             if r6.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r6.success()
             else:
-                error = r6.json().get("error", "???")
-                errcode = r6.json().get("errcode", "???")
+                error = r6.js.get("error", "???")
+                errcode = r6.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
             # print("OPRF success - Got response: ", json.dumps(r6.js, indent=4))
@@ -1002,17 +587,17 @@ class LocustClient(Client):
         }
         # with self.client.request("POST", url, headers=headers, json=body, catch_response=True) as r3:
         with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r7:
-            completed = r7.json().get("completed", [])
+            completed = r7.js.get("completed", [])
             if r7.status_code != 200:
-                error = r7.json().get("error", "???")
-                errcode = r7.json().get("errcode", "???")
+                error = r7.js.get("error", "???")
+                errcode = r7.js.get("errcode", "???")
                 print("Got error response: %s %s" % (errcode, error))
-            print("Register success - Got response: ", json.dumps(r7.json(), indent=4))
+            print("Register success - Got response: ", json.dumps(r7.js, indent=4))
 
-            self.user_id = r7.json().get("user_id", None)
-            self.access_token = r7.json().get("access_token", None)
+            self.user_id = r7.js.get("user_id", None)
+            self.access_token = r7.js.get("access_token", None)
             self.matrix_domain = self.user_id.split(":")[-1]
-            self.device_id = r7.json().get("device_id", None)
+            self.device_id = r7.js.get("device_id", None)
 
     def login_uia(self) -> None:
         """TODO: Update to make this a generic UIA handler that calls callbacks depending on stages
@@ -1032,18 +617,16 @@ class LocustClient(Client):
 
         # Request 1: Empty #####################################################
         body = {"identifier": {"type": "m.id.user", "user": self.user_id}}
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json=body, catch_response=True
-        ) as r1:
-            initial_json = r1.json()
-            session_id = r1.json().get("session", None)
+        with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r1:
+            initial_json = r1.js
+            session_id = r1.js.get("session", None)
 
             # print("Got response: ", json.dumps(r1.js, indent=4))
             if r1.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r1.success()
             else:
-                error = r1.json().get("error", "???")
-                errcode = r1.json().get("errcode", "???")
+                error = r1.js.get("error", "???")
+                errcode = r1.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
 
@@ -1063,17 +646,15 @@ class LocustClient(Client):
             },
         }
         r2_params = None
-        with self.locust_user.client.request(
-            "POST", url, headers=headers, json=body, catch_response=True
-        ) as r2:
-            completed = r2.json().get("completed", [])
-            r2_params = r2.json().get("params", {})
+        with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r2:
+            completed = r2.js.get("completed", [])
+            r2_params = r2.js.get("params", {})
 
             if r2.status_code == HTTPStatus.UNAUTHORIZED:  # 401
                 r2.success()
             else:
-                error = r2.json().get("error", "???")
-                errcode = r2.json().get("errcode", "???")
+                error = r2.js.get("error", "???")
+                errcode = r2.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
 
@@ -1104,18 +685,18 @@ class LocustClient(Client):
         }
         with self.locust_user.client.request("POST", url, headers=headers, json=body, catch_response=True) as r3:
             # with self.client.request("POST", url, headers=headers, json=body, catch_response=True) as r3:
-            completed = r3.json().get("completed", [])
+            completed = r3.js.get("completed", [])
             if r3.status_code != 200:
-                error = r3.json().get("error", "???")
-                errcode = r3.json().get("errcode", "???")
+                error = r3.js.get("error", "???")
+                errcode = r3.js.get("errcode", "???")
                 print(f"Got error response: {errcode} {error}")
                 return
-            print("Login success - Got response: ", json.dumps(r3.json(), indent=4))
+            print("Login success - Got response: ", json.dumps(r3.js, indent=4))
 
-            self.user_id = r3.json().get("user_id", None)
-            self.access_token = r3.json().get("access_token", None)
+            self.user_id = r3.js.get("user_id", None)
+            self.access_token = r3.js.get("access_token", None)
             self.matrix_domain = self.user_id.split(":")[-1]
-            self.device_id = r3.json().get("device_id", None)
+            self.device_id = r3.js.get("device_id", None)
 
     @logged_in
     def room_send(
